+++ std-bits/table/src/main/java/org/enso/table/parsing/problems/AdditionalInvalidRows.java
@@ -1,4 +1,6 @@
 package org.enso.table.parsing.problems;
 
+import org.enso.table.problems.Problem;
+
 /** A problem which indicates how many additional invalid rows were encountered. */
-public record AdditionalInvalidRows(long count) implements ParsingProblem {}
+public record AdditionalInvalidRows(long count) implements Problem {}